minor fixes

+++ scripts/compile-lang.sh
@@ -1,13 +1,17 @@
 #!/usr/bin/env bash
 
-echo "Copying files in src/ to src-working/ ..."
+echo "Copying files in src to src-working ..."
 rm -rf src-working
 cp -r src src-working
 
 node scripts/compile-lang.js $1 $2 $3 $4 $5 $6 $7
+if [ $? -ne 0 ]; then
+  echo "Custom build script failed :("
+  exit 1
+fi
 
 echo 'Invoking webpack ...'
 webpack --config webpack.config.js --entry ./src-working/index.js
 
-echo "Removing src-working/"
+echo "Removing src-working"
 rm -r src-working
